Make vagrant easy to use
by moving the Vagrantfile in the repository root and merging all vms
into one file. Vagrant vms can now be created without needing to switch
into a subdirectory anymore.

Also update the documentation to explain how to use vagrant and the test
inventory to include the new machines for CentOS 7, 8 and Ubuntu 18.04.